feat: Move test files under bddtests folder
closes #6
